Rename files that have CSV
